Make SearchResultsInfo.numMatching() return a long

+++ src/main/java/co/caio/tablier/model/SearchResultsInfo.java
@@ -12,7 +12,7 @@ public interface SearchResultsInfo {
 
   int paginationEnd();
 
-  int numMatching();
+  long numMatching();
 
   @Value.Default
   default String previousPageHref() {
